extract miscellaneous LocalBGS operations Move miscellaneous operations out so that: - it is easier to look at base-level functionality without all of these various operation implementations intermingled - these various operation implementations can eventually be organized according to the standards we've used for other operations
