+++ modules/claims_api/lib/bgs_service/claim_management_service.rb
@@ -24,5 +24,9 @@ module ClaimsApi
 
       make_request(endpoint: bean_name, action: 'updateClaimLevelSuspense', body:)
     end
+
+    def to_camelcase(claim:)
+      claim.deep_transform_keys { |k| k.to_s.camelize(:lower) }
+    end
   end
 end
